Address review comments - Removed UI elements from AnalogIOInterface, now hard-coded to be in Volts, with a range of +/- 10 Volts - Fixed OutputClock register writes, now they correctly set parameters chosen in the settings interface - Added manual update to the settings of specific devices. Note: This infrastructure will be modified at a later date to be less brittle - Unified the location of the enable button across tabs - Removed additional context resets that were leading to slow-downs in reading frames

+++ Source/Devices/OutputClock.cpp
@@ -38,7 +38,7 @@ bool OutputClock::updateSettings()
 	oni_reg_val_t baseFreqHz = deviceContext->readRegister(deviceIdx, (oni_reg_addr_t)OutputClockRegisters::BASE_FREQ_HZ);
 
 	auto periodCycles = (uint32_t)(baseFreqHz / frequencyHz);
-	auto h = (uint32_t)(periodCycles * (dutyCycle / 100));
+	auto h = (uint32_t)(periodCycles * ((double)dutyCycle / 100.0));
 	auto l = periodCycles - h;
 	auto delayCycles = (uint32_t)(delay * baseFreqHz);
